Added HandyDraw2, an alternative implementation.
Rather than using an offscreen buffer (which does not smooth properly), it intercepts all calls and passed them either to HandyRender or the canvas itself.

A disadvantage is that all public methods have to be overridden which will cause problems if these change in later implementations of PGraphics. We could choose to omit methods that we don't want to support - these will have no effect.